refactor the code

+++ backend/internal/models/user.go
@@ -22,9 +22,9 @@ type User struct {
 	FatherName    string             `bson:"father_name,omitempty" json:"father_name,omitempty"`
 	AadharNumber  string             `bson:"aadhar_number,omitempty" json:"aadhar_number,omitempty"`
 	TenthSchool   string             `bson:"tenth_school,omitempty" json:"tenth_school,omitempty"`
-	TenthMarks    int                `bson:"tenth_marks,omitempty" json:"tenth_marks,omitempty"`
+	TenthMarks    float64            `bson:"tenth_marks,omitempty" json:"tenth_marks,omitempty"`
 	TwelfthSchool string             `bson:"twelfth_school,omitempty" json:"twelfth_school,omitempty"`
-	TwelfthMarks  int                `bson:"twelfth_marks,omitempty" json:"twelfth_marks,omitempty"`
+	TwelfthMarks  float64            `bson:"twelfth_marks,omitempty" json:"twelfth_marks,omitempty"`
 	Cutoff        int                `bson:"cutoff,omitempty" json:"cutoff,omitempty"`
 	IsActive      bool               `bson:"is_active" json:"is_active"`
 	IsApproved    bool               `bson:"is_approved" json:"is_approved"`
